app/requirements.txt
python-ffmpeg==2.0.12
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
python-dotenv==1.0.1

azure-storage-blob==12.24.0
azure-identity==1.19.0

isort==5.13.2
autoflake==2.3.1
